requirements.txt
f5-icontrol-rest
six
json
